fix(w0-06): clear the lint findings this wave introduced
Three raw #fff literals in the period control. The rule that forbids them exists
because a constant colour renders identically in both themes, so any component
using one is permanently wrong in whichever theme it was not written for. They
are var(--text-inverse) now, which is what the DataTable's own active chips use.

isPermissionError moved out of WidgetState.tsx into widgetError.ts. Exporting a
function beside a component breaks fast refresh, and the classification is
wanted by callers that render their own shell anyway.

And a real dependency bug: `trend?.points ?? []` allocates a fresh array on every
render, so the series useMemo below it re-ran every time regardless of whether
the data had changed. Memoised on `trend`.

What is left in src/features/dashboard is pre-existing and stays visible: three
errors in shared.tsx (verified by stashing this wave and re-linting) and one
exhaustive-deps warning in the War Room effect, which predates this branch.
